CRS-2439: Suppress SDS40 hint text for ERS30 defaulting
When the ERSED has been defaulted to the ERS30 commencement date we should not display the SDS40 hint text "40/50% date has been applied".

Fixed a couple of tests runners that did not apply the feature toggles as per the input json.
