+++ src/main/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/service/CalculationResultEnrichmentService.kt
@@ -270,7 +270,7 @@ class CalculationResultEnrichmentService(
   private fun sds40Hint(type: ReleaseDateType, calculationBreakdown: CalculationBreakdown): ReleaseDateHint? {
     if (calculationBreakdown.breakdownByReleaseDateType.containsKey(type)) {
       val rules = calculationBreakdown.breakdownByReleaseDateType[type]!!.rules
-      if (rules.contains(CalculationRule.SDS_EARLY_RELEASE_APPLIES) && type.isEarlyReleaseHintType && !rules.contains(CalculationRule.HDCED_ADJUSTED_TO_365_COMMENCEMENT)) {
+      if (rules.contains(CalculationRule.SDS_EARLY_RELEASE_APPLIES) && type.isEarlyReleaseHintType && !rules.contains(CalculationRule.HDCED_ADJUSTED_TO_365_COMMENCEMENT) && !rules.contains(CalculationRule.ERSED_ADJUSTED_TO_ERS30_COMMENCEMENT)) {
         return ReleaseDateHint("40% date has been applied")
       }
       if ((
@@ -284,7 +284,7 @@ class CalculationResultEnrichmentService(
         val trancheText = if (rules.contains(CalculationRule.SDS_EARLY_RELEASE_ADJUSTED_TO_TRANCHE_1_COMMENCEMENT)) "1" else "2"
         return ReleaseDateHint("Defaulted to tranche $trancheText commencement")
       }
-      if (rules.contains(CalculationRule.SDS_STANDARD_RELEASE_APPLIES) && type.isStandardReleaseHintType && !rules.contains(CalculationRule.HDCED_ADJUSTED_TO_365_COMMENCEMENT)) {
+      if (rules.contains(CalculationRule.SDS_STANDARD_RELEASE_APPLIES) && type.isStandardReleaseHintType && !rules.contains(CalculationRule.HDCED_ADJUSTED_TO_365_COMMENCEMENT) && !rules.contains(CalculationRule.ERSED_ADJUSTED_TO_ERS30_COMMENCEMENT)) {
         return ReleaseDateHint("50% date has been applied")
       }
     }
