+++ src/test/resources/test_data/hint-text/expected-results/crs-2439-ac1-hint-text.json
@@ -1,0 +1,22 @@
+[
+  {
+    "type": "SLED",
+    "date": "2030-08-16",
+    "hints": []
+  },
+  {
+    "type": "CRD",
+    "date": "2027-04-11",
+    "hints": ["Friday, 09 April 2027 when adjusted to a working day", "40% date has been applied"]
+  },
+  {
+    "type": "ERSED",
+    "date": "2025-09-23",
+    "hints": []
+  },
+  {
+    "type": "ESED",
+    "date": "2032-01-25",
+    "hints": []
+  }
+]
