+++ src/test/resources/test_data/hint-text/expected-results/crs-2439-ac2-hint-text.json
@@ -1,0 +1,27 @@
+[
+  {
+    "type": "SLED",
+    "date": "2028-10-31",
+    "hints": []
+  },
+  {
+    "type": "CRD",
+    "date": "2025-08-19",
+    "hints": ["40% date has been applied"]
+  },
+  {
+    "type": "HDCED",
+    "date": "2025-02-21",
+    "hints": ["40% date has been applied"]
+  },
+  {
+    "type": "ERSED",
+    "date": "2024-10-22",
+    "hints": ["Defaulted to tranche 2 commencement"]
+  },
+  {
+    "type": "ESED",
+    "date": "2029-10-31",
+    "hints": []
+  }
+]
