+++ src/test/resources/test_data/hint-text/expected-results/crs-2439-ac3-hint-text.json
@@ -1,0 +1,27 @@
+[
+  {
+    "type": "SLED",
+    "date": "2030-10-30",
+    "hints": []
+  },
+  {
+    "type": "CRD",
+    "date": "2027-02-28",
+    "hints": ["Friday, 26 February 2027 when adjusted to a working day", "40% date has been applied"]
+  },
+  {
+    "type": "HDCED",
+    "date": "2026-03-01",
+    "hints": ["Monday, 02 March 2026 when adjusted to a working day", "40% date has been applied"]
+  },
+  {
+    "type": "ERSED",
+    "date": "2025-09-02",
+    "hints": ["40% date has been applied"]
+  },
+  {
+    "type": "ESED",
+    "date": "2031-03-13",
+    "hints": []
+  }
+]
