+++ src/test/resources/test_data/hint-text/expected-results/crs-2439-ac4-hint-text.json
@@ -1,0 +1,27 @@
+[
+  {
+    "type": "SLED",
+    "date": "2030-10-02",
+    "hints": []
+  },
+  {
+    "type": "CRD",
+    "date": "2027-06-15",
+    "hints": ["40% date has been applied"]
+  },
+  {
+    "type": "HDCED",
+    "date": "2026-06-16",
+    "hints": ["40% date has been applied"]
+  },
+  {
+    "type": "ERSED",
+    "date": "2025-12-01",
+    "hints": ["40% date has been applied"]
+  },
+  {
+    "type": "ESED",
+    "date": "2024-06-18",
+    "hints": []
+  }
+]
