+++ src/test/resources/test_data/hint-text/expected-results/crs-2439-ac5-hint-text.json
@@ -1,0 +1,27 @@
+[
+  {
+    "type": "SLED",
+    "date": "2029-04-15",
+    "hints": []
+  },
+  {
+    "type": "CRD",
+    "date": "2025-09-08",
+    "hints": ["40% date has been applied"]
+  },
+  {
+    "type": "HDCED",
+    "date": "2025-03-13",
+    "hints": ["40% date has been applied"]
+  },
+  {
+    "type": "ERSED",
+    "date": "2024-10-18",
+    "hints": ["50% date has been applied"]
+  },
+  {
+    "type": "ESED",
+    "date": "2029-04-15",
+    "hints": []
+  }
+]
