+++ src/test/resources/test_data/hint-text/input-data/crs-2439-ac3-hint-text.json
@@ -1,0 +1,175 @@
+{
+  "booking": {
+    "offender": {
+      "reference": "A1234EZ",
+      "dateOfBirth": "1996-01-01",
+      "isActiveSexOffender": false
+    },
+    "sentences": [
+      {
+        "type": "StandardSentence",
+        "offence": {
+          "committedAt": "2023-05-14",
+          "offenceCode": "TH68023"
+        },
+        "duration": {
+          "durationElements": {
+            "DAYS": 0,
+            "WEEKS": 0,
+            "YEARS": 0,
+            "MONTHS": 81
+          }
+        },
+        "isSDSPlus": false,
+        "identifier": "730c9170-1ad9-327b-af44-c517fb22a06b",
+        "recallType": null,
+        "sentencedAt": "2023-12-14",
+        "caseSequence": 1,
+        "lineSequence": 1,
+        "caseReference": "C1",
+        "externalSentenceId": {
+          "sentenceSequence": 1
+        },
+        "consecutiveSentenceUUIDs": [],
+        "isSDSPlusOffenceInPeriod": true,
+        "hasAnSDSEarlyReleaseExclusion": "VIOLENT",
+        "isSDSPlusEligibleSentenceTypeLengthAndOffence": false
+      },
+      {
+        "type": "StandardSentence",
+        "offence": {
+          "committedAt": "2023-08-01",
+          "offenceCode": "ID10001"
+        },
+        "duration": {
+          "durationElements": {
+            "DAYS": 0,
+            "WEEKS": 0,
+            "YEARS": 0,
+            "MONTHS": 6
+          }
+        },
+        "isSDSPlus": false,
+        "identifier": "a3556c6d-0b1f-3ccd-b18b-7597391b2ab1",
+        "recallType": null,
+        "sentencedAt": "2023-12-14",
+        "caseSequence": 1,
+        "lineSequence": 2,
+        "caseReference": "C1",
+        "externalSentenceId": {
+          "sentenceSequence": 2
+        },
+        "consecutiveSentenceUUIDs": [
+          "730c9170-1ad9-327b-af44-c517fb22a06b"
+        ],
+        "isSDSPlusOffenceInPeriod": false,
+        "hasAnSDSEarlyReleaseExclusion": "NO",
+        "isSDSPlusEligibleSentenceTypeLengthAndOffence": false
+      },
+      {
+        "type": "StandardSentence",
+        "offence": {
+          "committedAt": "2023-08-01",
+          "offenceCode": "ID10001"
+        },
+        "duration": {
+          "durationElements": {
+            "DAYS": 0,
+            "WEEKS": 0,
+            "YEARS": 0,
+            "MONTHS": 12
+          }
+        },
+        "isSDSPlus": false,
+        "identifier": "7ee43b9d-68ed-3f24-a17c-c7f24974cee0",
+        "recallType": null,
+        "sentencedAt": "2023-12-14",
+        "caseSequence": 1,
+        "lineSequence": 3,
+        "caseReference": "C1",
+        "externalSentenceId": {
+          "sentenceSequence": 9
+        },
+        "consecutiveSentenceUUIDs": [],
+        "isSDSPlusOffenceInPeriod": false,
+        "hasAnSDSEarlyReleaseExclusion": "NO",
+        "isSDSPlusEligibleSentenceTypeLengthAndOffence": false
+      },
+      {
+        "type": "StandardSentence",
+        "offence": {
+          "committedAt": "2023-08-01",
+          "offenceCode": "PC02021C"
+        },
+        "duration": {
+          "durationElements": {
+            "DAYS": 0,
+            "WEEKS": 0,
+            "YEARS": 0,
+            "MONTHS": 12
+          }
+        },
+        "isSDSPlus": false,
+        "identifier": "7ee43b9d-68ed-3f24-a17c-c7f24974cee0",
+        "recallType": null,
+        "sentencedAt": "2023-12-14",
+        "caseSequence": 1,
+        "lineSequence": 3,
+        "caseReference": "C1",
+        "externalSentenceId": {
+          "sentenceSequence": 9
+        },
+        "consecutiveSentenceUUIDs": [],
+        "isSDSPlusOffenceInPeriod": false,
+        "hasAnSDSEarlyReleaseExclusion": "NO",
+        "isSDSPlusEligibleSentenceTypeLengthAndOffence": false
+      },
+      {
+        "type": "StandardSentence",
+        "offence": {
+          "committedAt": "2023-04-23",
+          "offenceCode": "TH68023"
+        },
+        "duration": {
+          "durationElements": {
+            "DAYS": 0,
+            "WEEKS": 0,
+            "YEARS": 0,
+            "MONTHS": 67
+          }
+        },
+        "isSDSPlus": false,
+        "identifier": "ed20ec03-0420-33ac-bc18-29fa85e09aaa",
+        "recallType": null,
+        "sentencedAt": "2023-12-14",
+        "caseSequence": 1,
+        "lineSequence": 4,
+        "caseReference": "C1",
+        "externalSentenceId": {
+          "sentenceSequence": 16
+        },
+        "consecutiveSentenceUUIDs": [],
+        "isSDSPlusOffenceInPeriod": true,
+        "hasAnSDSEarlyReleaseExclusion": "VIOLENT",
+        "isSDSPlusEligibleSentenceTypeLengthAndOffence": false
+      }
+    ],
+    "adjustments": {
+      "REMAND": [
+        {
+          "toDate": "2023-12-13",
+          "fromDate": "2023-08-02",
+          "numberOfDays": 134,
+          "appliesToSentencesFrom": "2023-12-14"
+        }
+      ]
+    },
+    "externalMovements": [],
+    "historicalTusedData": null,
+    "returnToCustodyDate": null,
+    "fixedTermRecallDetails": null
+  },
+  "userInputs": {
+    "calculateErsed": true
+  }
+}
